cleaned up namespaces and added way to adjust bounds
